requirements.txt
appnope==0.1.4
asttokens==3.0.0
blinker==1.9.0
click==8.1.8
comm==0.2.2
debugpy==1.8.12
decorator==5.1.1
executing==2.2.0
filelock==3.17.0
Flask==3.1.0
fsspec==2025.2.0
ipykernel==6.29.5
ipython==8.32.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.5
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numpy==2.2.3
packaging==24.2
parso==0.8.4
pexpect==4.9.0
platformdirs==4.3.6
prompt_toolkit==3.0.50
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.1
PySastrawi==1.2.0
python-dateutil==2.9.0.post0
pyzmq==26.2.1
regex==2024.11.6
setuptools==75.8.0
six==1.17.0
stack-data==0.6.3
sympy==1.13.1
torch==2.6.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.12.2
wcwidth==0.2.13
Werkzeug==3.1.3
